+++ tests/test_event.py
@@ -86,6 +86,7 @@ def test_from_dict_simple(simple_event_data: DictFromJSON) -> None:
         "…)"
     )
     assert my_event.url == "https://spond.com/client/sponds/E1/"
+    assert my_event.is_cancelled is False
 
 
 def test_from_dict_additional_fields(complex_event_data: DictFromJSON) -> None: